探索処理の改善

+++ src/info/keloud/leJOS/utils/Arm.java
@@ -17,7 +17,7 @@ public class Arm extends AbstractUtil {
         centerMotor.setSpeed(speed);
     }
 
-    public void run(int speed, int angle) {
+    public void run(float speed, int angle) {
         setSpeed(speed);
         setAngle(angle);
         run();
