+++ src/info/keloud/leJOS/utils/ForwardWithColor.java
@@ -12,13 +12,13 @@ public class ForwardWithColor extends AbstractUtil {
         this.colorSensor = colorSensor;
     }
 
-    public void run(int speed, int colorId) {
+    public void run(float speed, int colorId) {
         setSpeed(speed);
         setColorId(colorId);
         run();
     }
 
-    public void run(int speed, String colorId) {
+    public void run(float speed, String colorId) {
         setSpeed(speed);
         setColorId(colorId);
         run();
@@ -29,7 +29,7 @@ public class ForwardWithColor extends AbstractUtil {
         // 初期化
         setOperationMode("Forward Color");
         int initTachoCount = leftMotor.getTachoCount();
-        int speedMin = 100;
+        float speedMin = 100;
         int degreeTachoCount = 0;
         leftMotor.setSpeed(speedMin);
         rightMotor.setSpeed(speedMin);
